pyproject.toml
[project]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Programming Language :: Python",
    "Topic :: Documentation",
    "Topic :: Software Development",
    "Topic :: Text Processing",
    "Typing :: Typed",
]
dependencies = [
    "anyconfig",
    "beautifulsoup4",
    "checksumdir",
    "commonmark",
    "cookiecutter",
    "copier",
    "deepmerge",
    "diskcache",
    "feedgen",
    "jinja2",
    "linkify-it-py",
    "markdown-it-py[plugins]",
    "markdown2[all]",
    "more-itertools",
    "pathspec",
    "pillow",
    "pluggy",
    "polyfactory",
    "pydantic>=2.0",
    "pydantic_extra_types>=2.0",
    "pydantic_settings",
    "pymdown-extensions",
    "python-frontmatter",
    "python-slugify",
    "pytz",
    "rich",
    "toml",
    "typer",
    'dateparser',
]

dynamic = ["version"]
description = "Static site generator plugins all the way down."
keywords = ["static-site"]
name = "markata"
readme = "README.md"
requires-python = ">=3.6"

[[project.authors]]
name = "Waylon Walker"
email = "waylon@waylonwalker.com"

[[project.authors]]
email = "waylon@markata.dev"

[build-system]
requires = ["hatchling>=1.4.1"]
build-backend = "hatchling.build"

[project.license]
file = "LICENSE"

[project.optional-dependencies]
tui = ["textual", "trogon"]
dev = [
    "black",
    "hatch",
    "interrogate",
    "mypy",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-tmp-files",
    "ruff",
]
pyinstrument = ["pyinstrument"]

[project.urls]
Homepage = "https://markata.dev"
Changelog = "https://markata.dev/CHANGELOG/"
Source = "https://github.com/waylonwalker/markata"
Issues = "https://github.com/waylonwalker/markata/issues"
Documentation = "https://markata.dev"

[project.scripts]
markata = "markata.cli:cli"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_untyped_calls = true
allow_untyped_globals = false
ignore_missing_imports = true
[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "markata.background.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "diskcache.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "frontmatter.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "bs4.*"
ignore_errors = true

[tool.hatch.version]
path = "markata/__about__.py"

[tool.coverage.run]
branch = true
parallel = true
omit = ["markata/__about__.py"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.hatch.envs.default]
dependencies = [
    "isort",
    "black",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-tmp-files",
    "ruff",
    "pyinstrument",
]

[tool.hatch.envs.dev]
dependencies = [
    "black",
    "ipython",
    "polyfactory",
    "black==22.10.0",
    "ipython",
    "pytest",
    "pytest-cov",
    "pytest-tmp-files",
    "ruff",
]

[tool.hatch.envs.test]
[[tool.hatch.envs.test.matrix]]
python = ["38", "39", "310", "311"]

[tool.hatch.envs.default.scripts]
cov = "pytest tests --cov-report=term-missing --cov-config=pyproject.toml --cov=markata --cov=tests"
no-cov = "cov --no-cov"
lint = "ruff markata"
format = "black --check markata"
build-docs = "markata build"
lint-format = ['lint', 'format']
lint-test = ["lint", "format", "cov"]
test-lint = "lint-test"
build = 'markata build'
serve = "python -m http.server --bind 0.0.0.0 8000 --directory markout"
ruff-fix = "ruff --fix markata"
black-fix = "black markata"
fix = ["black-fix", "ruff-fix"]

[tool.hatch.envs.test.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=markata --cov=tests"

[tool.hatch.build.targets.wheel]

[tool.hatch.build.targets.sdist]
exclude = ["/.github"]

[tool.ruff]
# E501 let black control line length errors
ignore = ["E501"]

target-version = "py37"
[tool.ruff.per-file-ignores]
'tests/**/*.py' = ['S101']
select = ["A", "E"]
